+++ kwcoco/channel_spec.py
@@ -5,8 +5,8 @@ image. This could be as simple as ``red|green|blue``, or more complex like:
 
 This functionality has been moved to "delayed_image".
 """
-from delayed_image.channel_spec import FusedChannelSpec
-from delayed_image.channel_spec import ChannelSpec
+from delayed_image.channel_spec import FusedChannelSpec  # NOQA
+from delayed_image.channel_spec import ChannelSpec  # NOQA
 import ubelt as ub
 ub.schedule_deprecation(
     'kwcoco', 'kwcoco.channel_spec', 'module', 'use delayed_image.channel_spec instead',